+++ EntradaSalida/Perifericos/LEDs/CLASE_LedWS2812.h
@@ -3,10 +3,10 @@
     class LedWS2812: public LedRGBW {
         private:
             Adafruit_NeoPixel *tiraRGBW;
-            unsigned short indiceTira;
+            size_t indiceTira;
 
         public:
-            LedWS2812(ColorRGBW colorApagado, ColorRGBW colorEncendido, Adafruit_NeoPixel *tiraRGBW, unsigned short indiceTira, bool estadoInicial = false);
+            LedWS2812(ColorRGBW colorApagado, ColorRGBW colorEncendido, Adafruit_NeoPixel *tiraRGBW, size_t indiceTira, bool estadoInicial = false);
             
             virtual void mostrar() override;
     };
